Finished updating tests for new requirement
Decided on definition of Lowest Common Ancestor for Directed Acyclic
Graphs. The definition of the lowest single common ancestor is:
the LCSA of two nodes u,v is the node L which lies on all root-u and
root-v paths, but no descendant of L has this property.

The new/updated tests test for this requirement.